Separated controllers to individual classes.
